fix: extract Android specific safe area logic and simplify safe area components/hooks
